\name{hire.data}
\alias{hire.data}
\docType{data}
\title{
The Data for the Hiring Problem
}
\description{
A matrix with two columns each giving the ranks  assigned by two mangers
to the \code{hire.questionaire$design} cases.  The managers choose first
by university, then by Sex (the first chooses Male
over Female, the second Female over Male), then
by dress.  Hair is ignored.
}
\usage{data(hire.data)}
\format{
  A matrix of ranks with 12 rows and two columns
}

\examples{
data(hire.data)
}
\keyword{datasets}
