\name{hire.despack}
\alias{hire.despack}
\docType{data}
\title{
A Design Pack for the Hire Problem
}
\description{
This design pack contains 26 9 card designs
and the corresponding line numbers from
\code{hire.questionaires$design}.
}
\usage{data(hire.despack)}
\format{
  The format is:\cr
List of 2\cr
 $ samps  :List of 26 vectors of 12 integres\cr
 $ designs:List of 26 data frames, each 9 obs. of  4 variables\cr
 }
\examples{
data(hire.despack)
}
\keyword{datasets}
