\name{hire.questionaire}
\alias{hire.questionaire}
\docType{data}
\title{
A Design for the Hiring Problem
}
\description{
This is a Conjoint Analysis design with 12 cards
}
\usage{data(hire.questionaire)}
\format{
  The format is:
List of 3\cr
 $ base.design:'data.frame':	9 obs. or 4 variables\cr 
 $ added      :'data.frame':	3 obs. of 4 variables\cr
 $ design     :'data.frame':	12 obs. of  4 variables:
}

\examples{
data(hire.questionaire)
}
\keyword{datasets}
